% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescription.R
\name{dropFields}
\alias{dropFields}
\title{Drop fields from a data.frame object}
\usage{
dropFields(df, fieldsToBeDropped)
}
\arguments{
\item{df}{a data.frame object}

\item{fieldsToBeDropped}{a vector of field names to be dropped}
}
\value{
the updated data.frame object
}
\description{
Drop fields from a data.frame object
}
