% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialFunctions.R
\docType{data}
\name{exampleRecruitment}
\alias{exampleRecruitment}
\title{An example of recruitment occurrence modelling.}
\format{
An object of class \code{data.frame} with 26248 rows and 7 columns.
}
\usage{
data(exampleRecruitment)
}
\description{
An example of recruitment occurrence modelling.
}
\examples{
data(exampleRecruitment)
}
\keyword{datasets}
