% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescription.R
\name{fillMissingValues}
\alias{fillMissingValues}
\title{Fill missing values in a dataset.}
\usage{
fillMissingValues(
  df,
  missingNumeric = -99999999,
  missingInteger = as.integer(missingNumeric),
  missingCharacter = "NA"
)
}
\arguments{
\item{df}{a data.frame object}

\item{missingNumeric}{a double that stands for missing value}

\item{missingInteger}{an integer that stands for missing value}

\item{missingCharacter}{a character string that stands for missing value}
}
\value{
an updated data.frame instance
}
\description{
Fill missing values in a dataset.
}
