% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmerGaussHermiteQuadrature.R
\docType{data}
\name{fitGLMER}
\alias{fitGLMER}
\title{An example of logistic regression with mixed effects.}
\format{
An object of class \code{glmerMod} of length 1.
}
\usage{
data(fitGLMER)
}
\description{
An example of logistic regression with mixed effects.
}
\examples{
data(fitGLMER)
}
\keyword{datasets}
