% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussHermiteQuadrature.R
\name{getHermiteQuadratureApproximation}
\alias{getHermiteQuadratureApproximation}
\title{Implementation of the Gauss-Hermite quadrature}
\usage{
getHermiteQuadratureApproximation(f, s2, ...)
}
\arguments{
\item{f}{a function whose first argument is a point of the quadrature}

\item{s2}{the variance of the variable}

\item{...}{other arguments to be passed to the function}
}
\description{
Implementation of the Gauss-Hermite quadrature
}
\author{
Mathieu Fortin - January 2024


Compute Gauss-Hermite quadrature.

The function relies on a five-point quadrature. It assumes the variable that follows the quadrature is
normally distribution with mean 0 and variance s2.
}
