% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialFunctions.R
\name{getPredictedProbsAndObservedProps}
\alias{getPredictedProbsAndObservedProps}
\title{Provides a table of predicted probabilities and observed proportions}
\usage{
getPredictedProbsAndObservedProps(
  dataSet,
  fitted,
  fieldName,
  classRange,
  obsFieldName,
  cutoff = 5
)
}
\arguments{
\item{dataSet}{a data.frame object that was used to fit the model}

\item{fitted}{a vector of predictions on the original scale}

\item{fieldName}{a character standing for the field name of the continuous explanatory variable in the dataSet object}

\item{classRange}{a numeric standing for the width of the classes}

\item{obsFieldName}{a character standing for the field name of the response variable}

\item{cutoff}{the minimum number of observations to consider the residual}
}
\value{
a data.frame object
}
\description{
Provides a table of predicted probabilities and observed proportions
}
