% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModelAssessmentFunctions.R
\name{getResidPlotAgainst}
\alias{getResidPlotAgainst}
\title{Plot average predicted values against classes of a particular
continuous covariate.}
\usage{
getResidPlotAgainst(
  model,
  data,
  fieldname,
  classWidth = NULL,
  printout = F,
  cutoff = NULL,
  normalize = T
)
}
\arguments{
\item{model}{a model object (e.g. an lme, gls, or lm instance) or a vector of residuals}

\item{data}{the data.frame object used to fit the model}

\item{fieldname}{the name of the variables the average residuals are plotted
against. By default, it is set to "pred" that is the model predictions}

\item{classWidth}{a positive numeric defining the width of the classes of
the variable on the x axis}

\item{printout}{a logical (true to print out the average residuals)}

\item{cutoff}{minimal number of observations for the average residual
to be included in the plot}

\item{normalize}{a logical true to use normalized residuals or false to
use raw residuals}
}
\description{
Plot average predicted values against classes of a particular
continuous covariate.
}
