% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmerGaussHermiteQuadrature.R
\name{glmmPA}
\alias{glmmPA}
\title{Approximate the population-average predictions from a
mixed-effects logistic regression  using Gauss-Hermite quadrature.}
\usage{
glmmPA(merObject, alpha = 0.05)
}
\arguments{
\item{merObject}{a glmerMod instance as produced by the glmer function}

\item{alpha}{the probability of type I error (alpha, by default set to 0.05)}
}
\value{
a data.frame object
}
\description{
The current implementation only supports a single random effect
on the intercept.
}
