% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialFunctions.R
\name{hosmerLemeshowTest}
\alias{hosmerLemeshowTest}
\title{Hosmer-Lemeshow test}
\usage{
hosmerLemeshowTest(observed, predicted, nbGroups = 10, covar = NULL)
}
\arguments{
\item{observed}{a vector of observed values}

\item{predicted}{a vector of predictions}

\item{nbGroups}{number of groups for the test (by default is set to 10)}

\item{covar}{an optional covariate to define the groups}
}
\value{
a list object
}
\description{
Hosmer-Lemeshow test
}
