% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescription.R
\name{renameField}
\alias{renameField}
\title{Rename a particular field of a data.frame object}
\usage{
renameField(df, oldName, newName)
}
\arguments{
\item{df}{a data.frame object}

\item{oldName}{the current field name}

\item{newName}{the new field name}
}
\value{
the updated data.frame object
}
\description{
Rename a particular field of a data.frame object
}
