% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescription.R
\name{summarizeThisVariable}
\alias{summarizeThisVariable}
\title{Provide a summary of a particular field}
\usage{
summarizeThisVariable(dataset, variable, onlyWherePresent = F)
}
\arguments{
\item{dataset}{a data.frame object}

\item{variable}{the field name (charactar)}

\item{onlyWherePresent}{a logical, if true the summary is calculated only for the
observations where the value is greater than 0 (by default is set to false)}
}
\description{
Provide a summary of a particular field
}
