% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{symSquare}
\alias{symSquare}
\title{Compute a vector from a symmetric matrix}
\usage{
symSquare(matrix)
}
\arguments{
\item{matrix}{a Matrix object}
}
\value{
a vector of values
}
\description{
Compute a vector from a symmetric matrix
}
