% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescription.R
\name{tabulateDataFrame}
\alias{tabulateDataFrame}
\title{Tabulate the data into a close-to-LateX format}
\usage{
tabulateDataFrame(dataFrame, nbDigits = NULL, prefix = NULL, suffix = NULL)
}
\arguments{
\item{dataFrame}{the data.frame object to be tabulated}

\item{nbDigits}{the number of numeric fields}

\item{prefix}{a prefix to field values}

\item{suffix}{a suffix to field values}
}
\description{
Tabulate the data into a close-to-LateX format
}
