% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialFunctions.R
\name{binomialResidualsClass}
\alias{binomialResidualsClass}
\title{Provides a graph with the binomial residuals}
\usage{
binomialResidualsClass(
  dataSet,
  fitted,
  fieldName,
  obsFieldName,
  min = -5,
  max = +5,
  cutoff = 5,
  title = NULL
)
}
\arguments{
\item{dataSet}{a data.frame object that was used to fit the model}

\item{fitted}{a vector of predictions on the original scale}

\item{fieldName}{a character standing for the field name of the categorical explanatory variable in the dataSet object}

\item{obsFieldName}{a character standing for the field name of the response variable}

\item{min}{the lower bound of the y-axis}

\item{max}{the upper bound of the y-axis}

\item{cutoff}{the minimum number of observations to consider the residual}

\item{title}{an optional character to be the title of the graph}
}
\value{
a ggplot2 graph
}
\description{
Binomial residuals are computed by classes of a categorical explanatory variable. The function
returns a ggplot2 graph.
}
