% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialFunctions.R
\name{binomialResidualsContinuous}
\alias{binomialResidualsContinuous}
\title{Provides a graph with the binomial residuals}
\usage{
binomialResidualsContinuous(
  dataSet,
  fitted,
  fieldName,
  classRange,
  obsFieldName,
  min = -5,
  max = +5,
  cutoff = 5,
  title = NULL,
  print = T,
  textsize = 18,
  xLabel = NULL
)
}
\arguments{
\item{dataSet}{a data.frame object that was used to fit the model}

\item{fitted}{a vector of predictions on the original scale}

\item{fieldName}{a character standing for the field name of the continuous explanatory variable in the dataSet object}

\item{classRange}{a numeric standing for the width of the classes}

\item{obsFieldName}{a character standing for the field name of the response variable}

\item{min}{the lower bound of the y-axis}

\item{max}{the upper bound of the y-axis}

\item{cutoff}{the minimum number of observations to consider the residual}

\item{title}{an optional character to be the title of the graph}

\item{print}{a logical (true prints the dataset before plotting it)}

\item{textsize}{the font size of the graph}

\item{xLabel}{a character string for the label of the x axis}
}
\value{
a ggplot2 graph
}
\description{
Binomial residuals are computed by classes of a continuous explanatory variable. The function
returns a ggplot2 graph.
}
