% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModelAssessmentFunctions.R
\docType{data}
\name{dataHeight}
\alias{dataHeight}
\title{Linear model assessment tools}
\format{
An object of class \code{data.frame} with 1060 rows and 6 columns.
}
\usage{
dataHeight
}
\description{
Linear model assessment tools
}
\author{
Mathieu Fortin - May 2025
}
\keyword{datasets}
